backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=backend/requirements.txt backend/requirements.in
#
alembic==1.13.1
    # via -r backend/requirements.in
amqp==5.3.1
    # via kombu
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.1.2
    # via
    #   -r backend/requirements.in
    #   passlib
billiard==4.2.1
    # via celery
black==23.12.1
    # via -r backend/requirements.in
celery==5.3.4
    # via -r backend/requirements.in
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
click==8.2.1
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   loguru
    #   pytest
    #   uvicorn
coverage[toml]==7.9.2
    # via pytest-cov
cryptography==45.0.5
    # via python-jose
deprecated==1.2.18
    # via limits
distlib==0.3.9
    # via virtualenv
ecdsa==0.19.1
    # via python-jose
fastapi==0.109.0
    # via
    #   -r backend/requirements.in
    #   sentry-sdk
filelock==3.18.0
    # via virtualenv
flake8==7.0.0
    # via -r backend/requirements.in
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.25.2
    # via -r backend/requirements.in
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
kombu==5.5.4
    # via celery
limits==5.4.0
    # via slowapi
loguru==0.7.2
    # via -r backend/requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
mccabe==0.7.0
    # via flake8
mypy==1.8.0
    # via -r backend/requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   kombu
    #   limits
    #   pytest
passlib[bcrypt]==1.7.4
    # via -r backend/requirements.in
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via
    #   black
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==3.6.0
    # via -r backend/requirements.in
prompt-toolkit==3.0.51
    # via click-repl
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.11.1
    # via flake8
pycparser==2.22
    # via cffi
pydantic==1.10.13
    # via
    #   -r backend/requirements.in
    #   fastapi
pyflakes==3.2.0
    # via flake8
pytest==7.4.4
    # via
    #   -r backend/requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.23.2
    # via -r backend/requirements.in
pytest-cov==4.1.0
    # via -r backend/requirements.in
python-dateutil==2.8.2
    # via
    #   -r backend/requirements.in
    #   celery
python-dotenv==1.0.0
    # via
    #   -r backend/requirements.in
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via -r backend/requirements.in
python-multipart==0.0.6
    # via -r backend/requirements.in
pyyaml==6.0.2
    # via
    #   pre-commit
    #   uvicorn
redis==5.0.1
    # via -r backend/requirements.in
rsa==4.9.1
    # via python-jose
sentry-sdk[fastapi]==1.39.0
    # via -r backend/requirements.in
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
slowapi==0.1.9
    # via -r backend/requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.25
    # via
    #   -r backend/requirements.in
    #   alembic
starlette==0.35.1
    # via fastapi
typing-extensions==4.14.1
    # via
    #   alembic
    #   fastapi
    #   limits
    #   mypy
    #   pydantic
    #   sqlalchemy
tzdata==2025.2
    # via
    #   celery
    #   kombu
urllib3==2.5.0
    # via sentry-sdk
uvicorn[standard]==0.27.0
    # via -r backend/requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.31.2
    # via pre-commit
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via uvicorn
win32-setctime==1.2.0
    # via loguru
wrapt==1.17.2
    # via deprecated
